use double for distance and elevationGain
change column names in trails table to distance_meters and elevation_gain_meters

+++ TrailFinder.Api.Tests/Controllers/TrailsControllerTests.cs
@@ -39,8 +39,8 @@ public class TrailsControllerTests
         
         string slug = "",
         string description = "",
-        decimal distance = 0,
-        decimal elevationGain = 0,
+        double distance = 0,
+        double elevationGain = 0,
         
         DifficultyLevel difficultyLevel = DifficultyLevel.Unknown,
         RouteType routeType = RouteType.Unknown,
